kernel 6.12 fix
Kernel 6.12 tty.operations.write api wants ssize_t and size_t

+++ xmm7360.c
@@ -1278,11 +1278,11 @@ static void xmm7360_tty_close(struct tty_struct *tty, struct file *filp)
 		tty_port_close(&qp->port, tty, filp);
 }
 
-static int xmm7360_tty_write(struct tty_struct *tty,
-			     const unsigned char *buffer, int count)
+static ssize_t xmm7360_tty_write(struct tty_struct *tty,
+				 const unsigned char *buffer, size_t count)
 {
 	struct queue_pair *qp = tty->driver_data;
-	int written;
+	ssize_t written;
 	written = xmm7360_qp_write(qp, buffer, count);
 	if (written < count)
 		qp->tty_needs_wake = 1;
